# SCS-Docs/api-reference/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "SCS API Documentation",
        "description": "SCS API Documentation for ECSE321 Group Project 2024.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/classTypes/{className}/approved/{isApproved}": {
            "put": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Change the approved status of a class type",
                "description": "Changes the approved status (true/false) of a class type",
                "operationId": "changeClassTypeApprovedStatus_1",
                "parameters": [
                    {
                        "name": "className",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isApproved",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful update of class type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassType"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input for updating class type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class type not found with the specified name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/classTypes/{className}": {
            "put": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Update class type by name",
                "description": "Updates the class type with the specified name",
                "operationId": "updateClassTypeDescription_1",
                "parameters": [
                    {
                        "name": "className",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClassTypeRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful update of class type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassTypeResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input for updating class type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class type not found with the specified name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/classType": {
            "post": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Create a class type",
                "description": "Creates a new class type",
                "operationId": "createClassType",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClassTypeRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Class type created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassTypeResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/classTypes/approved/{isApproved}": {
            "get": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Get all class types by approved status",
                "description": "Retrieves all class types by approved status (true/false)",
                "operationId": "getAllClassTypesByIsApproved",
                "parameters": [
                    {
                        "name": "isApproved",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of class types",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassTypeListDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/classType/{className}": {
            "get": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Get class type by name",
                "description": "Retrieves the class type with the specified name",
                "operationId": "getClassType_1",
                "parameters": [
                    {
                        "name": "className",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of class type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassTypeResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class type not found with the specified name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Delete class type by name",
                "description": "Deletes the class type with the specified name",
                "operationId": "deleteClassType_1",
                "parameters": [
                    {
                        "name": "className",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful deletion of class type"
                    },
                    "404": {
                        "description": "Class type not found with the specified name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorDto": {
                "type": "object",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ClassTypeResponseDto": {
                "type": "object",
                "properties": {
                    "className": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "isApproved": {
                        "type": "boolean"
                    }
                }
            },
...
            "ClassType": {
                "type": "object",
                "properties": {
                    "className": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "isApproved": {
                        "type": "boolean"
                    }
                }
            },
            "ClassTypeRequestDto": {
                "type": "object",
                "properties": {
                    "className": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "isApproved": {
                        "type": "boolean"
                    }
                }
            },
...
            "ClassTypeListDto": {
                "type": "object",
                "properties": {
                    "classTypes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ClassTypeResponseDto"
                        }
                    }
                }
            }
        }
    }
}